+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Changelog for `ta4j`, roughly following [keepachangelog.com](http://keepachangel
 - :tada: **Enhancement** Added DayOfWeekRule for specifying days of the week to trade.
 - :tada: **Enhancement** Added TimeRangeRule for trading within time ranges.
 - :tada: **Enhancement** Added floor() and ceil() to Num.class
+- :tada: **Enhancement** Added getters getLow() and getUp() in CrossedDownIndicatorRule
 
 ## 0.13 (released November 5, 2019)